implementation of private subjects and public vars in service

+++ src/app/components/movie-card/movie-card.component.ts
@@ -7,6 +7,7 @@ import { MathRoundPipe } from '../../pipes/mathRound/math-round.pipe';
 import { ReduceStringPipe } from '../../pipes/reduceString/reduce-string.pipe';
 import { RouterLink } from '@angular/router';
 import { MovieService } from '../../services/movie/movie.service';
+import { Movie } from '../../models/movie.model';
 
 export class MyApplicationModule {}
 
@@ -29,27 +30,38 @@ export class MovieCardComponent {
   public STATIC_IMAGE_PATH = 'https://image.tmdb.org/t/p/w500/';
   public isInFavoriteList: boolean | null = false;
 
-  @Input() movie: any;
+  @Input() movie: Movie | null = null;
   @Input() isShortDescriptionNeeded = true;
 
   constructor(private movieService: MovieService) {}
 
   // Funcs for favorites
   setToFavoriteMovieList() {
-    this.movieService.setMovieToFavoriteMovieList(this.movie.id);
+    if (this.movie) {
+      this.movie.addedToFavoriteList = true;
+
+      this.movieService.setMovieToFavoriteMovieList(this.movie);
+    }
   }
   deleteMovieFromFavoriteMovieList() {
-    this.movie.addedToFavoriteList = false;
-    this.movieService.deleteMovieFromFavoriteMovieList(this.movie);
+    if (this.movie) {
+      this.movie.addedToFavoriteList = false;
+
+      this.movieService.deleteMovieFromFavoriteMovieList(this.movie);
+    }
   }
 
   // Func for watch later
   setToWatchLaterMovieList() {
-    this.movie.addedToWatchLaterList = true;
-    this.movieService.setToWatchLaterMovieList(this.movie);
+    if (this.movie) {
+      this.movie.addedToWatchLaterList = true;
+      this.movieService.setToWatchLaterMovieList(this.movie);
+    }
   }
   deleteMovieWatchLaterMovieList() {
-    this.movie.addedToWatchLaterList = false;
-    this.movieService.deleteMovieWatchLaterMovieList(this.movie);
+    if (this.movie) {
+      this.movie.addedToWatchLaterList = false;
+      this.movieService.deleteMovieWatchLaterMovieList(this.movie);
+    }
   }
 }
